refactor: 替换max/min为std::max/std::min并优化代码结构
- 将数学函数max/min替换为标准库std::max/std::min
- 修复类型转换警告，使用static_cast进行显式转换
- 优化初始化列表顺序，提高代码可读性
- 添加必要的头文件包含
- 移除过时的版本检查代码
- 调整构建配置，添加缺失的系统库链接
- 修复宏定义在非MSVC编译器的兼容性问题
- 优化容器操作，使用reserve预分配内存

+++ Easy2D/src/Animation/Animation.cpp
@@ -69,7 +69,7 @@ void easy2d::Animation::_update()
 		_last += _seq->getInterval();
 		++_frameIndex;
 
-		if (_frameIndex == frames.size())
+		if (static_cast<size_t>(_frameIndex) == frames.size())
 		{
 			this->stop();
 			break;
